engine: Enable OSK mode
Currently gnome-shell sends IBUS_CAP_OSK when OSK(On-Screen Keyboard)
is enabled in GNOME Wayland only.
However users still can type characters using the physical keyboard
while OSK is enabled so the current status is not perfect.

But ibus-anthy now commit the preedit when a candidate is clicked
on IBus lookup table instead of the selection in the OSK mode.

BUG=https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2415
BUG=https://gitlab.gnome.org/GNOME/mutter/-/issues/2441
BUG=rhbz#2015149